% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg.r
\name{aggreg}
\alias{aggreg}
\title{Function aggreg}
\usage{
aggreg(x, y = NULL, fun = NULL, verbose = T)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{y}{Vector or matrix specifying the subsets to compute summary statistics for.}

\item{fun}{Function specifying the summary statistic to compute. If NULL (default), mean will be calculated.}

\item{verbose}{Logical value specifying the verbocity of output.}
}
\description{
Splits a numeric vector or matrix into subsets, computes summary statistics for each, and returns the result in a convenient form.
}
\details{
Splits a numeric vector or matrix into subsets, computes summary statistics for each, and returns the result in a convenient form.
}
\examples{
aggreg(x=1:10,y=c(rep(1,5),rep(2,5)))
}
\keyword{modeling}
